requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
dlib==19.24.1
flatbuffers==23.5.8
fonttools==4.39.3
gast==0.4.0
gaze-tracking==0.0.1
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
jax==0.4.9
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
MouseInfo==0.1.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sqlparse==0.4.4
tensorboard-data-server==0.7.0
tensorflow-addons==0.20.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
typeguard==2.13.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
wrapt==1.14.1
